requirements.txt
asgiref==3.8.1
crispy-bootstrap5==2024.2
Django==5.0.6
django-appconf==1.0.6
django-bootstrap5==24.2
django-crispy-forms==2.2
django-htmx==1.18.0
django-imagekit==5.0.0
pilkit==3.0
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
sqlparse==0.5.0
typing_extensions==4.12.2
